LPS-30441 Move static logic from EditEntryAction to TrashUtil
